Разработан базовый механизм отправки смс через API SMSAero, созданы необходимые json-объекты с информацией.
