+++ kalendar.py
@@ -1,10 +1,11 @@
 #! python3
 # -*- coding: utf-8 -*-
 
-events_list = {(2018, 6, 17): ('Victor Osipov', 'Katya Osipova')}
-
-preachers_list = {'Victor Osipov':('+79129760922'),
-                  'Katya Osipova': ('+79965574689')}
-
-text = "Try to send text in Russian"
-# You shall not see this'
+events_2018_summer = [
+    [2018, 7, 21], [2018, 7, 22],
+    [2018, 7, 28], [2018, 7, 29],
+    [2018, 8, 4], [2018, 8, 5],
+    [2018, 8, 11], [2018, 8, 12],
+    [2018, 8, 18], [2018, 8, 19],
+    [2018, 8, 25], [2018, 8, 26]
+]
